Minor corrections in NetworkFrontend, Protocol, RemotePlayer and Scheduler. The connection of the first player seems to work. Having trouble when testing with other players (definitely because on the same computer) and whith the Protocol::writeQuint32 method (perhaps linked with the players troubles).

+++ src/BelotusCommon/Protocol.cpp
@@ -82,7 +82,7 @@ void Protocol::MessageProcessed()
 void Protocol::writeQuint32(const quint32 value)
 {
     qDebug() << "Protocol : writeQuint32 : " << value ;
-    *(this->out) << value;
+    *(this->out) << (quint32) value;
     qDebug() << "Protocol : Fin writeQuint32" ;
 }
